- DLL библиотеки вшиты в exe - Автоматическое определение языка при формировании настроек

+++ SP-EFT-ProfileEditor/Character.cs
@@ -149,7 +149,7 @@ namespace SP_EFT_ProfileEditor
                 {
 
                     [JsonProperty("StackObjectsCount")]
-                    public int? StackObjectsCount { get; set; }
+                    public long? StackObjectsCount { get; set; }
 
                     /// <summary>
                     /// Found in raid?
